project_structure - Copy/requirements.txt
############################################################
# CORE
############################################################
numpy
pandas
scipy
python-dateutil

############################################################
# VISUALIZATION
############################################################
matplotlib
seaborn
plotly
altair
pillow

############################################################
# MACHINE LEARNING ALGORITHMS
############################################################
scikit-learn
xgboost
lightgbm
catboost
statsmodels

############################################################
# FEATURE ENGINEERING & ENCODERS
############################################################
category-encoders
feature-engine
sklearn-pandas
imbalanced-learn

############################################################
# DIMENSION REDUCTION / CLUSTERING
############################################################
umap-learn
hdbscan

############################################################
# STORAGE / SERIALIZATION
############################################################
pickle5
joblib
pyarrow
fastparquet
openpyxl
xlrd

############################################################
# DATA PROCESSING UTILITIES
############################################################
tqdm
pyyaml
requests
regex

############################################################
# STREAMLIT APP & UI
############################################################
streamlit
streamlit-option-menu
streamlit-aggrid

############################################################
# BACKEND / API (OPTIONAL)
############################################################
fastapi
uvicorn

############################################################
# DATABASE CONNECTORS (OPTIONAL)
############################################################
sqlalchemy
pymysql
psycopg2
sqlite3-binary

############################################################
# LOGGING & MONITORING
############################################################
loguru
rich
coloredlogs

############################################################
# MODEL EVALUATION & EXPLAINABILITY
############################################################
scikit-plot
yellowbrick
shap
lime
